Migrate contents of 'equations' component to 'equation-detail'
